new extended method for simple replies
