+++ src/commands/games-mp/gunfight.ts
@@ -41,11 +41,9 @@ export default class GunFightCommand extends Command {
 	 * @memberof GunFightCommand
 	 */
 	public async run(msg: KlasaMessage, [opponent]: [User]): Promise<KlasaMessage | KlasaMessage[]> {
-		if (opponent.bot) return msg.sendMessage('Bots may not be fought.', { reply: msg.author });
-
-		if (opponent.id === msg.author.id) return msg.sendMessage('You may not fight yourself.', { reply: msg.author });
-
-		if (this.fighting.has(msg.channel.id)) return msg.sendMessage('Only one fight may be occurring per channel.', { reply: msg.author });
+		if (opponent.bot) return msg.sendSimpleEmbedReply('Bots may not be fought.');
+		if (opponent.id === msg.author.id) return msg.sendSimpleEmbedReply('You may not fight yourself.');
+		if (this.fighting.has(msg.channel.id)) return msg.sendSimpleEmbedReply('Only one fight may be occurring per channel.');
 
 		this.fighting.add(msg.channel.id);
