+++ src/commands/games-mp/wizard-convention.ts
@@ -4,135 +4,147 @@
 
 import { stripIndents } from 'common-tags';
 import { Collection } from 'discord.js';
-import { awaitPlayers } from '@lib/helpers/base';
 import { Command, CommandStore, KlasaMessage } from 'klasa';
-import { questions, stories } from '../../extras/mafia';
+import * as data from '../../extras/wizard-convention';
+import { awaitPlayers } from '@lib/helpers/base';
 import { delay, shuffle } from '@lib/utils/util';
 
 /**
- * Starts a game of Mafia.
+ * Starts a game of Wizard Convention.
  *
  * @export
- * @class MafiaCommand
+ * @class WizardConventionCommand
  * @extends {Command}
  */
-export default class MafiaCommand extends Command {
+export default class WizardConventionCommand extends Command {
 
 	private playing: Set<string> = new Set();
 
 	/**
-	 * Creates an instance of MafiaCommand.
+	 * Creates an instance of WizardConventionCommand.
 	 *
 	 * @param {CommandoClient} client
-	 * @memberof MafiaCommand
+	 * @memberof WizardConventionCommand
 	 */
 	public constructor(store: CommandStore, file: string[], directory: string) {
 		super(store, file, directory, {
-			aliases: ['town-of-salem', 'werewolf'],
-			description: 'Who is the Mafia? Who is the doctor? Who is the detective? Will the Mafia kill them all?'
+			aliases: ['wiz-convention', 'wizard-convention'],
+			description: 'Who is the Dragon? Who is the healer? Who is the mind reader? Will the Dragon eat them all?'
 		});
+
 	}
 
 	/**
-	 * Run the "Mafia" command.
+	 * Run the "WizardConvention" command.
 	 *
 	 * @param {KlasaMessage} msg
 	 * @returns {(Promise<KlasaMessage | KlasaMessage[]>)}
-	 * @memberof MafiaCommand
+	 * @memberof WizardConventionCommand
 	 */
 	public async run(msg: KlasaMessage): Promise<KlasaMessage | KlasaMessage[]> {
-		if (this.playing.has(msg.channel.id)) return msg.sendMessage('Only one game may be occurring per channel.', { reply: msg.author });
+		if (this.playing.has(msg.channel.id)) return msg.sendSimpleEmbedReply('Only one game may be occurring per channel.');
 		this.playing.add(msg.channel.id);
+
 		try {
 			await msg.sendMessage('You will need at least 2 more players, at maximum 10. To join, type `join game`.');
-			const awaitedPlayers = await awaitPlayers(msg, 10, 3, { dmCheck: true });
+
+			const awaitedPlayers = await awaitPlayers(msg, 10, 3);
+
 			if (!awaitedPlayers) {
 				this.playing.delete(msg.channel.id);
 
 				return msg.sendMessage('Game could not be started...');
 			}
+
 			const players: any = await this.generatePlayers(awaitedPlayers);
 			let turn = 1;
-			while (players.size > 2 && players.some((p: any) => p.role === 'mafia')) {
-				let killed = null;
-				let saved = null;
+
+			while (players.size > 2 && players.some((p: any) => p.role === 'dragon')) {
+				let eaten = null;
+				let healed = null;
 
 				await msg.sendMessage(`Night ${turn}, sending DMs...`);
 
 				for (const player of players.values()) {
 					if (player.role.includes('pleb')) continue;
+
 					await msg.sendMessage(`The ${player.role} is making their decision...`);
+
 					const valid = players.filterArray((p: any) => p.role !== player.role);
+
 					await player.user.send(stripIndents`
-						${questions[player.role]} Please type the number.
+						${data.questions[player.role]} Please type the number.
 						${valid.map((p: any, i: number) => `**${i + 1}.** ${p.user.tag}`).join('\n')}
 					`);
+
 					const filter = (res: any): boolean => valid[Number.parseInt(res.content, 10) - 1];
 					const decision = await player.user.dmChannel.awaitMessages(filter, {
 						max: 1,
 						time: 120000
 					});
+
 					if (!decision.size) {
 						await player.user.send('Sorry, time is up!');
 						continue;
 					}
 					const choice = Number.parseInt(decision.first().content, 10);
-					if (player.role === 'mafia') {
+
+					if (player.role === 'dragon') {
 						const chosen = players.get(choice);
-						killed = chosen.id;
-						await player.user.send(`${chosen.user.tag} will be killed...`);
-					} else if (player.role === 'doctor') {
+						eaten = chosen.id;
+						await player.user.send(`${chosen.user.tag} will be eaten...`);
+					} else if (player.role === 'healer') {
 						const chosen = players.get(choice);
-						saved = chosen.id;
-						await player.user.send(`${chosen.user.tag} will be saved...`);
-					} else if (player.role === 'detective') {
-						await player.user.send(players.find((p: any) => p.role === 'mafia').id === choice ? 'Yes.' : 'No.');
+						healed = chosen.id;
+						await player.user.send(`${chosen.user.tag} will be healed...`);
+					} else if (player.role === 'mind reader') {
+						await player.user.send(players.find((p: any) => p.role === 'dragon').id === choice ? 'Yes.' : 'No.');
 					}
 				}
 
-				const display = killed ? players.get(killed).user : null;
-				const story = stories[Math.floor(Math.random() * stories.length)];
+				const display = eaten ? players.get(eaten).user : null;
+				const story = data.stories[Math.floor(Math.random() * data.stories.length)];
 
-				if (killed && killed !== saved) players.delete(killed);
+				if (eaten && eaten !== healed) players.delete(eaten);
 
-				if (killed && killed === saved) {
+				if (eaten && eaten === healed) {
 					await msg.sendMessage(stripIndents`
-						Late last night, a Mafia member emerged from the dark and tried to kill ${display}${story}
-						Thankfully, a doctor stepped in just in time to save the day.
-						Who is this mysterious Mafia member? You have one minute to decide.
+						Late last night, a dragon emerged and tried to eat ${display}${story}
+						Thankfully, a healer stepped in just in time to save the day.
+						Who is this mysterious dragon? You have one minute to decide.
 					`);
-				} else if (killed && players.size < 3) {
+				} else if (eaten && players.size < 3) {
 					await msg.sendMessage(stripIndents`
-						Late last night, a Mafia member emerged from the dark and killed poor ${display}${story}
-						Sadly, after the event, the final citizen left the town in fear, leaving the Mafia to rule forever.
+						Late last night, a dragon emerged and devoured poor ${display}${story}
+						Sadly, after the event, the final wizard ran in fear, leaving the dragon to rule forever.
 					`);
 					break;
-				} else if (killed && killed !== saved) {
+				} else if (eaten && eaten !== healed) {
 					await msg.sendMessage(stripIndents`
-						Late last night, a Mafia member emerged from the dark and killed poor ${display}${story}
-						Who is this mysterious Mafia member? You have one minute to decide.
+						Late last night, a dragon emerged and devoured poor ${display}${story}
+						Who is this mysterious dragon? You have one minute to decide.
 					`);
 				} else {
 					await msg.sendMessage(stripIndents`
-						Late last night, a Mafia member emerged from the dark. Thankfully, however, they didn't try to kill anyone.
-						Who is this mysterious Mafia member? You have one minute to decide.
+						Late last night, a dragon emerged. Thankfully, however, it didn't try to eat anyone.
+						Who is this mysterious dragon? You have one minute to decide.
 					`);
 				}
 
 				await delay(60000);
+
 				const playersArr = Array.from(players.values());
 
 				await msg.sendMessage(stripIndents`
-					Who do you think is the Mafia member? Please type the number.
+					Who do you think is the dragon? Please type the number.
 					${playersArr.map((p: any, i: number) => `**${i + 1}.** ${p.user.tag}`).join('\n')}
 				`);
 
-				const voted: any = [];
+				const voted: any[] = [];
+
 				const filter = (res: any): boolean => {
 					if (!players.exists((p: any) => p.user.id === res.author.id)) return false;
-
 					if (voted.includes(res.author.id)) return false;
-
 					if (!playersArr[Number.parseInt(res.content, 10) - 1]) return false;
 
 					voted.push(res.author.id);
@@ -144,31 +156,42 @@ export default class MafiaCommand extends Command {
 					max: players.size,
 					time: 120000
 				});
+
 				if (!votes.size) {
-					await msg.sendMessage('No one will be hanged.');
+					await msg.sendMessage('No one will be expelled.');
 					continue;
 				}
-				const hanged = this.getHanged(votes, players, playersArr);
-				await msg.sendMessage(`${hanged.user} will be hanged.`);
-				players.delete(hanged.id);
+
+				const expelled: any = this.getExpelled(votes, players, playersArr);
+
+				await msg.sendMessage(`${expelled.user} will be expelled.`);
+
+				players.delete(expelled.id);
+
 				++turn;
 			}
-			this.playing.delete(msg.channel.id);
-			const mafia = players.find((p: any) => p.role === 'mafia');
-			if (!mafia) return msg.sendMessage('The Mafia has been hanged! Thanks for playing!');
 
-			return msg.sendMessage(`Oh no, the Mafia wasn't caught in time... Nice job, ${mafia.user}!`);
+			this.playing.delete(msg.channel.id);
+
+			const dragon = players.find((p: any) => p.role === 'dragon');
+
+			if (!dragon) return msg.sendMessage('The dragon has been vanquished! Thanks for playing!');
+
+			return msg.sendMessage(`Oh no, the dragon wasn't caught in time... Nice job, ${dragon.user}!`);
 		} catch (err) {
 			this.playing.delete(msg.channel.id);
 
-			return msg.sendMessage(`Oh no, an error occurred: \`${err.message}\`. Try again later!`, { reply: msg.author });
+			return msg.sendSimpleError(`Oh no, an error occurred: \`${err.message}\`. Try again later!`);
 		}
 	}
 
 	private async generatePlayers(list: any): Promise<void> {
-		let roles = ['mafia', 'doctor', 'detective'];
+		let roles = ['dragon', 'healer', 'mind reader'];
+
 		for (let i = 0; i < (list.length - 2); i++) roles.push(`pleb ${i + 1}`);
+
 		roles = shuffle(roles);
+
 		const players = new Collection();
 		let i = 0;
 
@@ -183,11 +206,12 @@ export default class MafiaCommand extends Command {
 		}
 	}
 
-	private getHanged(votes: any, players: any, playersArr: any): any {
+	private getExpelled(votes: any, players: any, playersArr: any): any {
 		const counts: any = new Collection();
 
 		for (const vote of votes.values()) {
 			const player = players.get(playersArr[Number.parseInt(vote.content, 10) - 1].id);
+
 			if (counts.has(player.id)) {
 				++counts.get(player.id).votes;
 			} else {
