+++ src/commands/games-sp/chance.ts
@@ -37,9 +37,9 @@ export default class ChanceCommand extends Command {
 	 */
 	public async run(msg: KlasaMessage, [chance]: [number]): Promise<KlasaMessage | KlasaMessage[]> {
 		const loss = Math.floor(Math.random() * (chance ? Number(chance) : 1000));
-		if (!loss) return msg.sendMessage('Nice job! 10/10! You deserve some cake!', { reply: msg.author });
+		if (!loss) return msg.sendSimpleEmbedReply('Nailed it! You won!');
 
-		return msg.sendMessage('Nope, sorry, you lost.', { reply: msg.author });
+		return msg.sendSimpleEmbedReply('Sorry, you lost!');
 	}
 
 }
